MAILBOX-331 Move memory code related to quota mailing
