+++ mailbox/plugin/quota-mailing-memory/src/test/java/org/apache/james/mailbox/quota/memory/InMemoryQuotaMailingListenerTest.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.inmemory.quota;
+package org.apache.james.mailbox.quota.memory;
 
 import org.apache.james.mailbox.quota.mailing.QuotaMailingListenerTest;
 import org.junit.jupiter.api.extension.ExtendWith;
