+++ mailbox/plugin/quota-mailing-memory/src/test/java/org/apache/james/mailbox/quota/memory/InMemoryQuotaThresholdChangesStoreTest.java
@@ -17,23 +17,12 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.inmemory.quota;
+package org.apache.james.mailbox.quota.memory;
 
-import org.apache.james.mailbox.quota.QuotaThresholdChangesStore;
-import org.junit.jupiter.api.extension.ExtensionContext;
-import org.junit.jupiter.api.extension.ParameterContext;
-import org.junit.jupiter.api.extension.ParameterResolutionException;
-import org.junit.jupiter.api.extension.ParameterResolver;
+import org.apache.james.mailbox.quota.QuotaThresholdChangesStoreTest;
+import org.junit.jupiter.api.extension.ExtendWith;
 
-public class InMemoryQuotaThresholdChangesStoreExtension implements ParameterResolver {
+@ExtendWith(InMemoryQuotaThresholdChangesStoreExtension.class)
+public class InMemoryQuotaThresholdChangesStoreTest implements QuotaThresholdChangesStoreTest {
 
-    @Override
-    public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
-        return (parameterContext.getParameter().getType() == QuotaThresholdChangesStore.class);
-    }
-
-    @Override
-    public Object resolveParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
-        return new InMemoryQuotaThresholdChangesStore();
-    }
 }
