+++ mailbox/pom.xml
@@ -54,6 +54,7 @@
         <module>zoo-seq-provider</module>
 
         <module>plugin/quota-mailing</module>
+        <module>plugin/quota-mailing-memory</module>
         <module>plugin/spamassassin</module>
     </modules>
